:recycle: 화면 맨 하단 감지하는 함수 따로 파일로 분리
